# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.DepositResponse:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: number
    type: object
  schemas.DepositUpdateStock:
    properties:
      method:
        enum:
        - add
        - subtract
        - set
        example: add|subtract|set
        type: string
      product_id:
        type: integer
      stock:
        type: number
    type: object
...
  schemas.MovementStock:
    properties:
      amount:
        type: number
      from_id:
        type: integer
      from_type:
        enum:
        - deposit
        - point_sale
        type: string
      ignore_stock:
        type: boolean
      product_id:
        type: integer
      to_id:
        type: integer
      to_type:
        enum:
        - deposit
        - point_sale
        type: string
    required:
    - amount
    - from_id
    - product_id
    - to_id
    type: object
  schemas.MovementStockResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      from_id:
        type: integer
      from_type:
        type: string
      id:
        type: integer
      ignore_stock:
        type: boolean
      product:
        $ref: '#/definitions/schemas.ProductResponse'
      to_id:
        type: integer
      to_type:
        type: string
      user:
        $ref: '#/definitions/schemas.UserResponse'
    type: object
...
  schemas.PointSaleResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: number
    type: object
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  schemas.RoleResponse:
    properties:
      id:
        type: integer
      name:
        enum:
        - admin
        - vendedor
        - repositor
        type: string
    type: object
...
  schemas.UserResponse:
    properties:
      address:
        type: string
      cellphone:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      point_sales:
        items:
          $ref: '#/definitions/schemas.PointSaleResponse'
        type: array
      role:
        $ref: '#/definitions/schemas.RoleResponse'
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a API for Club Norte
  termsOfService: http://swagger.io/terms/
  title: APP Club Norte
  version: "1.0"
paths:
...
  /v1/deposit/get/{id}:
    get:
      consumes:
      - application/json
      description: DepositProductGetByID obtener un producto por ID del producto
      parameters:
      - description: Id del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.DepositResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: DepositProductGetByID
      tags:
      - Deposit
  /v1/deposit/get_all:
    get:
      consumes:
      - application/json
      description: DepositProductGetAll obtener productos por paginacion
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.DepositResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: DepositProductGetAll
      tags:
      - Deposit
  /v1/deposit/get_by_code:
    get:
      consumes:
      - application/json
      description: DepositProductGetByCode obtener un producto por codigo del deposito
      parameters:
      - description: codigo del producto
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.DepositResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: DepositProductGetByCode
      tags:
      - Deposit
  /v1/deposit/get_by_name:
    get:
      consumes:
      - application/json
      description: DepositProductGetByName obtener productos por por similitud de
        nombre
      parameters:
      - description: nombre por aproximacion del producto
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.DepositResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: DepositProductGetByName
      tags:
      - Deposit
  /v1/deposit/update_stock:
    put:
      consumes:
      - application/json
      description: DepositProductUpdateStock alctualizar stock de un producto del
        deposito
      parameters:
      - description: nombre por aproximacion del producto
        in: body
        name: stock_update
        required: true
        schema:
          $ref: '#/definitions/schemas.DepositUpdateStock'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.DepositResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: DepositProductUpdateStock
      tags:
      - Deposit
  /v1/movement_stock/get/{id}:
    get:
      consumes:
      - application/json
      description: MovementStockGet Obtener un movimiento de stock por ID
      parameters:
      - description: ID del movimiento de stock
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.MovementStockResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: MovementStockGet
      tags:
      - MovementStock
  /v1/movement_stock/get_all:
    get:
      consumes:
      - application/json
      description: MovementStockGetAll Obtener movimeintos de sotck por paginacion
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 10
        description: Número de elementos por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.MovementStockResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: MovementStockGetAll
      tags:
      - MovementStock
  /v1/movement_stock/move:
    post:
      consumes:
      - application/json
      description: MovementStock movimiento de stock entre doposito y puntos de ventas
      parameters:
      - description: movimiento de stock
        in: body
        name: movement_stock
        required: true
        schema:
          $ref: '#/definitions/schemas.MovementStock'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: MovementStock
      tags:
      - MovementStock
...
securityDefinitions:
  CookieAuth:
    description: 'Enter your JWT token here. Example: "eyJhbGciOiJIUz..."'
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"